Fix up comments
- better split of hash-parser/project-loader unit tests.
